palproj/app/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
langchain==0.2.13
langchain-google-genai==1.0.10
langchain-community==0.2.12
python-dotenv==1.0.0
requests
Flask-Session
pytest